travel agent management

+++ src/main/scala/ua/kpi/ipsa/domain/HotelStarCategory.scala
@@ -1,3 +1,3 @@
 package ua.kpi.ipsa.domain
 
-case class HotelStarCategory(id: Int, stars: Int, description: String, region: String)
+case class HotelStarCategory(id: Long, stars: Int, description: String, region: String)
